bugfix em components com descricao vazia e uso do apollo-client para query dos repositorios
