+++ src/pages/home/index.js
@@ -1,21 +1,18 @@
-import React, { useState, useEffect } from 'react'
+import React, { useState } from 'react'
 
-import { getRepositoriesSearchQuery } from '../../lib/github'
+import { getRepositoriesSearchQuery } from '../../utils/githubSearch'
 import { useQuery } from '@apollo/react-hooks'
 import repositoriesQuery from '../../graphql/repositories.graphql'
-import { SortField } from '../../lib/constants'
+import { SortField } from '../../utils/constants'
 import { RepositoryGrid, OptionBar } from '../../components'
 
 import './homePage.css'
 
 const HomePage = () => {
   const [sort, setSort] = useState(SortField.STARS_DESC)
-  const [cursor, setCursor] = useState(null)
   const { data, loading, error, fetchMore } = useQuery(repositoriesQuery, {
     variables: {
-      first: 12,
       repositories: getRepositoriesSearchQuery(sort),
-      cursor,
     },
   })
 
@@ -23,39 +20,25 @@ const HomePage = () => {
   const { nodes: repositories, pageInfo } = search || {}
   const { hasNextPage, endCursor } = pageInfo || {}
 
-  useEffect(() => {
-    if (endCursor) {
-      setCursor(endCursor.replace('=', ''))
-    }
-  })
-
-  const updateRepositories = () => {
+  const onFetchMore = () => {
     fetchMore({
       query: repositoriesQuery,
       variables: {
-        first: 12,
         repositories: getRepositoriesSearchQuery(sort),
-        cursor,
+        cursor: endCursor,
       },
       updateQuery: (previousResult, { fetchMoreResult }) => {
-        const previousEntry = previousResult.entry
-        console.log(previousEntry)
         const { search: newSearch } = fetchMoreResult || {}
-        const { nodes: newRepositories, pageInfo } = newSearch || {}
-        const { hasNextPage, endCursor: newCursor } = pageInfo || {}
-        if (endCursor) {
-          setCursor(endCursor.replace('=', ''))
-        }
+        const { nodes: newRepositories, pageInfo, repositoryCount } =
+          newSearch || {}
+
         return {
-          endCursor: newCursor,
-          hasNextPage,
-          entry: {
-            repositories: [
-              ...newRepositories,
-              ...previousEntry.data.repositories,
-            ],
+          search: {
+            __typename: previousResult.search.__typename,
+            pageInfo,
+            nodes: [...previousResult.search.nodes, ...newRepositories],
+            repositoryCount,
           },
-          __typename: previousEntry.__typename,
         }
       },
     })
@@ -72,10 +55,7 @@ const HomePage = () => {
       <RepositoryGrid repositories={repositories || []} />
       <div className="footer">
         {hasNextPage && (
-          <button
-            type="button"
-            className="show-more-btn"
-            onClick={updateRepositories}>
+          <button type="button" className="show-more-btn" onClick={onFetchMore}>
             {loading ? (
               <div className="loader" role="img" aria-label="loading" />
             ) : (
